Move `fr` unit to "CSS Flexible Lengths" (#129)
* Move `fr` unit to "CSS Flexible Lengths"

The link for `fr` on https://developer.mozilla.org/en-US/docs/Web/CSS/Reference points to the wrong page. This is because the `fr` unit doesn't belong with the other lengths, such as `em` or `px`. Rather, it is defined on the `<flex>` data type page.

We will also need to update https://github.com/mdn/kumascript/blob/master/macros/CSS_Ref.ejs.

* Add "CSS Flexible Lengths"