Fix <transform-function> syntax
Previously syntax describes one or more functions in any order repeated one or more times. However, according to spec it should be a regular enum. Repetition is defined in `transform-list`.
Although the old syntax is quite correct (can be used for a value matching), it is overwhelmed. This commit simplifies the syntax.
There is no `<transform-function>` syntax description in [CSS Transforms Module Level 1](https://www.w3.org/TR/css-transforms-1/#typedef-transform-function), but it can be generated by a description.