Change for percentage property for `inset`
`logicalHeightOrWidthOfContainingBlock` has been created in both 
https://github.com/mdn/data/blob/master/css/properties.schema.json 
and 
https://github.com/mdn/data/blob/master/l10n/css.json.

According to the documentation `inset` percentage value is relative to the containing block’s size in the corresponding axis (e.g. width for left or right, height for top and bottom)

Now on the page dedicated to this property, it says the following:
__Percentages	logical-height of containing block__
which is not entirely true.